+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+Code.load_file("test/utils.exs")
 ExUnit.start()
